% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhenMeasureChaosR-package.R
\docType{package}
\name{WhenMeasureChaosR-package}
\alias{WhenMeasureChaosR}
\alias{WhenMeasureChaosR-package}
\title{WhenMeasureChaosR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
